feat(parser): add recursive parser
Added a module to recursively import the content of the yaml / json file using `import_from` key. I was hoping to be able to do this directly in Pydantic  but it would have been really complicated. With this approach we will load the whole data once as a dictionary and then use `parse_obj` classmethod to create the object.