feat(artifacts): artifacts to pull from source and copy to temp folder
To provide more task independence it makes sense to force each task to copy any artifact from
somewhere into a temporary folder that is used for the duration of the task and then copy whatever
output artifacts are produced from the temp folder back to permanent location. This means a lot of
the commands don't need so many file path parameters to be templated. I also think it makes more
sense from a user perspective to "use this file" rather than "give the path to this file".