perf(pydantic): bump to pydantic 1.0.0
This version bump is done to enable the use of `root_validators` which are validators that apply to
the entire data class object rather than specific fields. Previously when field cross-validation was
made there the object returned in values was most often a partial version of the entire dataclass to
be deserialised.