feat(workflow): add workflow template value hydration
The workflow can no retrieve and replace wherever a user implements a template value of type {{
workflow.x.y }} with it's actual value in the workflow. It's worth bearing in mind that this will
return errors when the value in question is not set. This method is intended to be used mostly by
the workflow runner plugins to help template reading.