fix(pydantic): revert to v0.32.2
This has been implemented to Queenbee can be compatible with FastAPI for microservice development.
Had to change a couple of functions and add a manual `validate_all` function to the workflow and
function schemas. On a side note: thank the gods for automated tests written beforehand to ensure
similar behaviour!

#17