fix(packaging): generate tar files without using filesystem
Generating tar files that are output at BytesIO objects without needing to write any intermediary
resource to disk means we can manage tar on a server relatively easily. It also segregates the
responsibility of generating the file bytes from that of writing them (to a file, an S3 bucket or a
database).

closes #96