fix(artifacts_source): replace Enum with constr in type
This was causing an issue with the documentation: https://travis-ci.org/github/ladybug-tools/queenbee/jobs/708701110#L536

I believe this is a result of the recent change in how Pydantic works with Enum objects. We can use the Enum trick anymore. The good news is that `constr` is there to help.