refator(io): move default and required to DAGGenericInput classes
This is not an ideal change but was necessary to keep the dotnet code inheritance clean. Since the `default` type is different between the baseclass and the subclass it would result is using Queenbee baseclass instead.

With this change it should use the GenericInput class as the baseclass for all the inputs and inputs-alias classes.