feat(api version): add an api_version key to DTOs
Data Transfer Objects (DTOs) are any object that we expect to be used to be used by an external API
(ie: not-queenbee) to be parsed and computed. I have opted for a version and maturity flag for API
versions (ie: version = v1 and maturity = beta1). This versioning schema should ideally work like
described for the Kubernetes project:
https://kubernetes.io/docs/concepts/overview/kubernetes-api/#api-groups-and-versioning